// examples/export-data/utils/types.ts
export interface CustomGraphType {
    _id: string | Types.ObjectId;
    eventTypeId: string | Types.ObjectId;
    projectId: string | Types.ObjectId;
    graphTitle: string;
    xProperty: string;
    yProperty: string;
    graphType: string;
    caption?: string;
}